chore: updated code for remote config version 0.3.1

+++ appambit_sdk_react_native/android/src/main/java/com/appambit/AppambitRemoteConfigModule.kt
@@ -24,8 +24,8 @@ class AppambitRemoteConfigModule(reactContext: ReactApplicationContext) :
     return RemoteConfig.getBoolean(key)
   }
 
-  override fun getInt(key: String?): Double {
-    return RemoteConfig.getInt(key).toDouble()
+  override fun getLong(key: String?): Double {
+    return RemoteConfig.getLong(key).toDouble()
   }
 
   override fun getDouble(key: String?): Double {
