+++ appambit_sdk_react_native/src/NativeAppambitRemoteConfig.ts
@@ -4,7 +4,7 @@ export interface Spec extends TurboModule {
     enable(): void;
     getString(key: string): string;
     getBoolean(key: string): boolean;
-    getInt(key: string): number;
+    getLong(key: string): number;
     getDouble(key: string): number;
 }
